feat: 메뉴 선택 관련 책임 이전
